+++ changelog.md
@@ -1,5 +1,18 @@
 # WebMan Templates Changelog
 
+## 2.1.0
+
+* **Add**: Global templates
+* **Fix**: WebMan Amplifier plugin admin notification
+
+### Files changed:
+
+	changelog.md
+	webman-templates.php
+	includes/classes/class-webman-templates.php
+	templates/_global/rows.dat
+
+
 ## 2.0.1
 
 * **Update**: Icelander template files